requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
numpy==2.1.2
openpyxl==3.1.5
python-multipart==0.0.9
openai==1.42.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.11.7
pydantic-settings==2.11.0
starlette>=0.27.0
